Fix topic list display on Suggest a Conf w Firefox (#423)
As noted in the MDN and detailed in the (gulp) 17 year old bugzilla entry - https://bugzilla.mozilla.org/show_bug.cgi?id=40545, Firefox doesn't adequately display option elements with a label and no content. Instead the is a ~20px wide list of blank entries which is quite unuseable for Firefox users.